Update index.md
Hi, I have added a plugin on the page as I was unsure where or if it should go in the data/entries.js. It is intended for more generic video embeds so doesn't match up to a particular product. Please let me know if this plugin makes the grade and if there is any information you need. Thanks.